Clean-up in traits and input
