+++ src/traits.rs
@@ -1,8 +1,16 @@
-use input;
 use gust::mesh;
 use core::program;
 use core::surface;
 use core::texture;
+use glm;
+
+pub struct ReflectingInput {
+    pub model: glm::Matrix4<f32>,
+    pub view: glm::Matrix4<f32>,
+    pub projection: glm::Matrix4<f32>,
+    pub normal: glm::Matrix4<f32>,
+    pub camera_position: glm::Vec3
+}
 
 #[derive(Debug)]
 pub enum Error {
@@ -40,5 +48,5 @@ pub trait Emitting {
 }
 
 pub trait Reflecting {
-    fn reflect(&self, input: &input::ReflectingInput) -> Result<(), Error>;
+    fn reflect(&self, input: &ReflectingInput) -> Result<(), Error>;
 }
